Made table name box editable, eliminated Table Name popup
Made the table name box editable so that the table name can be defined
at the same time as the fields are entered. I eliminated the Table Name
popup box because it was now unnecessary. Table name validation was
moved to a separate function and is now called from addTableDialog.html
so that table names are still validated.