Made table name editable for existing tables, put addTableDialog in it's own js file
Made table name editable for existing tables. This required
modifications to the table panels (to fix edit and delete events as well
as the table name label). It also required changing the name of the
table in the tables array, and changing the ID of the table panel. Also
moved the addTableDialog javascript into it's own .js file. Being
located in the .html file made it harder to debug with Developer Tools
and performance of embedded js is typically slower.