Ability to erase canvas and fixed bugs when starting from empty canvas
Added ability to erase the canvas. This is done with a button that
clears the tables from local storage and from the canvas so that the
user can start over. Also fixed a bug where the tables object isn't
initialized right if the first time information box is displayed. To fix
this I moved the information box down to be the last thing that happens
after the window loads instead of the first.