Extracted common plumbing code into methods, pre-loaded template
Extracted common plumbing code into methods. Some pieces of the jsPlumb
connector code were duplicated (but not perfectly) and this was causing
some unpredictable behavior. Also preloaded the table template because
the asynchronous get() request was causing operations to happen in a
random order and data wasn't available when it was expected to be. This
was especially problematic when creating new tables.